requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
cryptography
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
alembic
